% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assignClasses}
\alias{assignClasses}
\title{Internal use only}
\usage{
assignClasses(df, inVars)
}
\arguments{
\item{df}{A data frame}

\item{inVars}{The variables expected in the df}
}
\value{
A data frame with corrected column classes
}
\description{
Use the variables file to assign classes to each column in each data file
}
\keyword{internal}
