% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R
\name{assignHorVer}
\alias{assignHorVer}
\alias{assignHorVer}
\title{For IS files, make two new columns for horizontal position (HOR) and vertical position (VER)}
\usage{
assignHorVer(df, splitName)

assignHorVer(df, splitName)
}
\arguments{
\item{df}{A data frame of data}

\item{splitName}{The name of the corresponding data file, split by period into a list of component parts}

\item{df}{A data frame of data}

\item{splitName}{The name of the corresponding data file, split by period into a list of component parts}
}
\value{
a data frame with the new columns

a data frame with the new columns
}
\description{
For IS files, make two new columns for horizontal position (HOR) and vertical position (VER)

For IS files, make two new columns for horizontal position (HOR) and vertical position (VER)
}
