% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUp.R
\name{cleanUp}
\alias{cleanUp}
\title{Clean up folder after stacking}
\usage{
cleanUp(folder)
}
\arguments{
\item{folder}{The file path to the folder that needs to be cleaned up (the root directory of the data package)}
}
\value{
Any existing directory except for stackedFiles will be removed. Nothing is returned.
}
\description{
Remove unzipped monthly data folders
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
