% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find.datatables}
\alias{find.datatables}
\title{List the names of the data tables within each folder}
\usage{
find.datatables(filepath, fnames = T)
}
\arguments{
\item{filepath}{The filepath of the input file}

\item{fnames}{Full names - if true, then return the full file names including enclosing folders, if false, return only the file names}
}
\value{
a data frame of file names
}
\description{
List the names of the data tables within each folder
}
