% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{find.tables.unique}
\alias{find.tables.unique}
\alias{find.tables.unique}
\title{Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)}
\usage{
find.tables.unique(datatables)

find.tables.unique(datatables)
}
\arguments{
\item{datatables}{A list of data files}

\item{datatables}{A list of data files}
}
\value{
An array of unique table names

An array of unique table names
}
\description{
Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)

Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)
}
