% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTablesUnique.R
\name{findTablesUnique}
\alias{findTablesUnique}
\title{Find unique data tables in dataset}
\usage{
findTablesUnique(datatables, tabletypes)
}
\arguments{
\item{datatables}{A list of data files}
}
\value{
An array of unique table names
}
\description{
Find the unique data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
