% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{get.filesize}
\alias{get.filesize}
\alias{get.filesize}
\title{Get the file's size in megabytes}
\usage{
get.filesize(filepath)

get.filesize(filepath)
}
\arguments{
\item{filepath}{The path to the file}

\item{filepath}{The path to the file}
}
\value{
The size of the file in megabytes

The size of the file in megabytes
}
\description{
Get the file's size in megabytes

Get the file's size in megabytes
}
