% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVariables}
\alias{getVariables}
\title{Internal use only}
\usage{
getVariables(varFile)
}
\arguments{
\item{varFile}{A file that contains variable definitions}
}
\value{
A data frame with fieldName and assigned column class, along with table if present
}
\description{
Support way to force R to read assign correct data types to each column based on variables file
}
\keyword{internal}
