% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{list.files.inZip}
\alias{list.files.inZip}
\alias{list.files.inZip}
\title{Given the top level zip file, return dataframe of all of the files within it without
unzipping the file}
\usage{
list.files.inZip(zippath)

list.files.inZip(zippath)
}
\arguments{
\item{zippath}{The path to a zip file}

\item{zippath}{The path to a zip file}
}
\value{
A list of filenames within the given zip file

A list of filenames within the given zip file
}
\description{
Given the top level zip file, return dataframe of all of the files within it without
unzipping the file

Given the top level zip file, return dataframe of all of the files within it without
unzipping the file
}
