% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{list.zipfiles}
\alias{list.zipfiles}
\alias{list.zipfiles}
\title{Given the data frame of all the files within the top level zip file,
return just an array of just the zip file names (no pdfs, etc.)}
\usage{
list.zipfiles(zippath)

list.zipfiles(zippath)
}
\arguments{
\item{zippath}{The path to a zip file}

\item{zippath}{The path to a zip file}
}
\value{
An array of all zip files contained within the focal zip file

An array of all zip files contained within the focal zip file
}
\description{
Given the data frame of all the files within the top level zip file,
return just an array of just the zip file names (no pdfs, etc.)

Given the data frame of all the files within the top level zip file,
return just an array of just the zip file names (no pdfs, etc.)
}
