% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFilesInZip.R
\name{listFilesInZip}
\alias{listFilesInZip}
\title{Get a data frame with the names of all files within a zipped NEON data package}
\usage{
listFilesInZip(zippath)
}
\arguments{
\item{zippath}{The path to a zip file}
}
\value{
A list of filenames within the given zip file
}
\description{
Given the top level zip file, return dataframe of all of the files within it without
unzipping the file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
