% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listZipfiles.R
\name{listZipfiles}
\alias{listZipfiles}
\title{Get all zip file names within a zipped NEON data package}
\usage{
listZipfiles(zippath)
}
\arguments{
\item{zippath}{The path to a zip file}
}
\value{
An array of all zip files contained within the focal zip file
}
\description{
Given the data frame of all the files within the top level zip file,
return an array of just the zip file names (no pdf, xml, or other files).
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
