% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{stackByTable}
\alias{stackByTable}
\alias{stackByTable}
\title{stackByTable: Join data files by table type.}
\usage{
stackByTable(filepath)

stackByTable(filepath)
}
\arguments{
\item{filepath}{The location of the zip file}

\item{filepath}{The location of the zip file}
}
\value{
All files are unzipped and one file for each table type is created and written.

All files are unzipped and one file for each table type is created and written.
}
\description{
Given a zipped data file, do a full join of all data files, grouped by table type.
This should result in a small number of large files.

Given a zipped data file, do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
