% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{stackDataFiles}
\alias{stackDataFiles}
\alias{stackDataFiles}
\title{Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.}
\usage{
stackDataFiles(folder)

stackDataFiles(folder)
}
\arguments{
\item{folder}{The location of the data}

\item{folder}{The location of the data}
}
\value{
One file for each table type is created and written.

One file for each table type is created and written.
}
\description{
Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.

Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
