% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stackfiles}
\alias{stackfiles}
\title{Bind all of the data from all of the monthly packages for each of the dataset's tables}
\usage{
stackfiles(tables, outpath)
}
\arguments{
\item{tables}{Array of table names returned from find.tables.unique()}

\item{outpath}{The location that files have been unpacked to}
}
\value{
A set of stacked files
}
\description{
Bind all of the data from all of the monthly packages for each of the dataset's tables
}
