% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFileToGeoCSV.R
\name{transformFileToGeoCSV}
\alias{transformFileToGeoCSV}
\title{Transform NEON CSV file to GeoCSV}
\usage{
transformFileToGeoCSV(infile, varfile, outfile)
}
\arguments{
\item{infile}{The path to the file that needs to be parsed}

\item{varfile}{The path to the variables file needed to parse the infile}

\item{outfile}{The path where the new GeoCSV file should be placed}
}
\value{
The same data file with a GeoCSV header
}
\description{
Read in a NEON monthly data zip file and parse the respective variables file to create a new GeoCSV file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
