% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{unzip.zipfile}
\alias{unzip.zipfile}
\alias{unzip.zipfile}
\title{Unzip a zip file either at just the top level or recursively through the file.}
\usage{
unzip.zipfile(zippath, outpath = substr(zippath, 1, nchar(zippath) - 4),
  level = "all")

unzip.zipfile(zippath, outpath = substr(zippath, 1, nchar(zippath) - 4),
  level = "all")
}
\arguments{
\item{zippath}{The filepath of the input file}

\item{outpath}{The name of the folder to save unpacked files to}

\item{level}{Whether the unzipping should occur only for the 'top' zip file, or unzip 'all' recursively}

\item{zippath}{The filepath of the input file}

\item{outpath}{The name of the folder to save unpacked files to}

\item{level}{Whether the unzipping should occur only for the 'top' zip file, or unzip 'all' recursively}
}
\description{
Unzip a zip file either at just the top level or recursively through the file.

Unzip a zip file either at just the top level or recursively through the file.
}
