% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{find.datatables}
\alias{find.datatables}
\alias{find.datatables}
\title{List the names of the data tables within each folder}
\usage{
find.datatables(folder, fnames = T)

find.datatables(folder, fnames = T)
}
\arguments{
\item{folder}{The folder of the outputs}

\item{fnames}{Full names - if true, then return the full file names including enclosing folders, if false, return only the file names}

\item{folder}{The folder of the outputs}

\item{fnames}{Full names - if true, then return the full file names including enclosing folders, if false, return only the file names}
}
\value{
a data frame of file names

a data frame of file names
}
\description{
List the names of the data tables within each folder

List the names of the data tables within each folder
}
