% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find.tables.unique}
\alias{find.tables.unique}
\title{Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)}
\usage{
find.tables.unique(filelist)
}
\arguments{
\item{filelist}{A list of data files}
}
\value{
a list of unique tables
}
\description{
Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)
}
