% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find.tables.unique}
\alias{find.tables.unique}
\title{Internal use only}
\usage{
find.tables.unique(datatables, tabletypes)
}
\arguments{
\item{datatables}{A list of data files}
}
\value{
An array of unique table names
}
\description{
Find the data tables that are present in the dataset (e.g., 2 minute vs 30 minute, or pinning vs identification data)
}
\keyword{internal}
