% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.filesize}
\alias{get.filesize}
\title{Get a file's size in megabytes}
\usage{
get.filesize(filepath)
}
\arguments{
\item{filepath}{The path to the file}
}
\value{
The size of the file in megabytes
}
\description{
Get a file's size in megabytes
}
