% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notes.R, R/functions.R
\name{getVariables}
\alias{getVariables}
\alias{getVariables}
\title{Support way to force R to read assign correct data types to each column based on variables file}
\usage{
getVariables(varFile)

getVariables(varFile)
}
\arguments{
\item{varFile}{A file that contains variable definitions}

\item{varFile}{A file that contains variable definitions}
}
\value{
A data frame with fieldName and assigned column class, along with table if present

A data frame with fieldName and assigned column class, along with table if present
}
\description{
Support way to force R to read assign correct data types to each column based on variables file

Support way to force R to read assign correct data types to each column based on variables file
}
