% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariables.R
\name{getVariables}
\alias{getVariables}
\title{Get correct data types}
\usage{
getVariables(varFile)
}
\arguments{
\item{varFile}{A file that contains variable definitions}
}
\value{
A data frame with fieldName and assigned column class, along with table if present
}
\description{
Support way to force R to read assign correct data types to each column based on variables file
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
\keyword{internal}
