% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list.files.inZip}
\alias{list.files.inZip}
\title{Given the top level zip file, return dataframe of all of the files within it without
unzipping the file}
\usage{
list.files.inZip(filepath)
}
\arguments{
\item{filepath}{The path to a zip file}
}
\value{
A list of filenames within the give zip file
}
\description{
Given the top level zip file, return dataframe of all of the files within it without
unzipping the file
}
