% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list.zipfiles}
\alias{list.zipfiles}
\title{Given the data frame of all the files within the top level zip file,
return just an array of just the zip file names (no pdfs, etc.)}
\usage{
list.zipfiles(filepath)
}
\arguments{
\item{filepath}{The path to a zip file}
}
\value{
An array any zip files contained within the focal zip file
}
\description{
Given the data frame of all the files within the top level zip file,
return just an array of just the zip file names (no pdfs, etc.)
}
