% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stackByTable}
\alias{stackByTable}
\title{Primary function: Join data files in a zipped NEON data package by table type}
\usage{
stackByTable(filepath, folder = F)
}
\arguments{
\item{filepath}{The location of the zip file}

\item{folder}{T or F: does the filepath point to a parent, unzipped folder, or a zip file? If F, assumes the filepath points to a zip file. Defaults to F.}
}
\value{
All files are unzipped and one file for each table type is created and written.
}
\description{
Given a zipped data file, do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
