% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stackByTable}
\alias{stackByTable}
\title{Unzip and stack the data in all data files, grouped by table type. This should result in a small number of large files.}
\usage{
stackByTable(location.package)
}
\arguments{
\item{filepath}{The location of the zip file}
}
\value{
All files are unzipped and one file for each table type is created and written.
}
\description{
Unzip and stack the data in all data files, grouped by table type. This should result in a small number of large files.
}
