% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackByTable.R
\name{stackByTable}
\alias{stackByTable}
\title{Join data files in a zipped NEON data package by table type}
\usage{
stackByTable(filepath, folder = FALSE)
}
\arguments{
\item{filepath}{The location of the zip file}

\item{folder}{T or F: does the filepath point to a parent, unzipped folder, or a zip file? If F, assumes the filepath points to a zip file. Defaults to F.}
}
\value{
All files are unzipped and one file for each table type is created and written.
}
\description{
Given a zipped data file, do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
Claire Lunch \email{clunch@battelleecology.org}
}
