% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackDataFiles.R
\name{stackDataFiles}
\alias{stackDataFiles}
\title{Join data files in a unzipped NEON data package by table type}
\usage{
stackDataFiles(folder)
}
\arguments{
\item{folder}{The location of the data}
}
\value{
One file for each table type is created and written.
}
\description{
Given a folder of unzipped files (unzipped NEON data file), do a full join of all data files, grouped by table type.
This should result in a small number of large files.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Christine Laney \email{claney@battelleecology.org}
}
